fix(main/toxic): enable openal and x11 to receive video and audio from calls
Fixes `video_call.c:58:16: error: no member named 'av' in 'struct Toxic'` during build when `libx11`, `libvpx` and `openal-soft` all happened to be installed in the prefix before compiling `toxic` (the manifestation of https://github.com/JFreegman/toxic/issues/483 in termux-packages)

This one is a bit of a far-out suggestion, but I noticed that audio **receiving** (openal playback) and video **receiving** (x11 render video stream) both **do** work if I actually install this build in Termux, open Termux:X11, open a terminal in Termux:X11, set up a Tox client on a desktop PC, add the client on Termux as a friend, start a call with it, set my microphone and webcam on the desktop PC client, turn on my camera on the desktop PC client, and watch and listen for the output on the Android device.

neither video **sending** (v4l2 recording) nor audio **sending** (openal recording) work, so the Termux:X11 client is only able to hear and see the other call participants and other participants can't hear or see the Termux:X11 client (though chat messages work), but I wondered whether that means that the X11 and openal support both have to remain completely disabled, or if it is worth it at all to enable them.